requirements.txt
astroid==3.3.5
coverage==7.6.3
dill==0.3.9
exceptiongroup==1.2.2
execnet==2.1.1
factory_boy==3.3.1
Faker==30.6.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
pylint==3.3.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-factoryboy==2.7.0
pytest-mock==3.14.0
pytest-vcr==1.0.2
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.16.0
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
vcrpy==6.0.2
wrapt==1.16.0
yarl==1.15.5
